+++ packages/core/src/layout/Sidebar/DefaultProviderSettings.tsx
@@ -26,7 +26,7 @@ import {
 } from '@backstage/core-api';
 import Star from '@material-ui/icons/Star';
 import React from 'react';
-import { OAuthProviderSettings, OIDCProviderSettings } from './Settings';
+import { ProviderSettingsItem } from './Settings';
 
 export const DefaultProviderSettings = () => {
   const configApi = useApi(configApiRef);
@@ -36,49 +36,49 @@ export const DefaultProviderSettings = () => {
   return (
     <>
       {providers.includes('google') && (
-        <OIDCProviderSettings
+        <ProviderSettingsItem
           title="Google"
           apiRef={googleAuthApiRef}
           icon={Star}
         />
       )}
       {providers.includes('microsoft') && (
-        <OIDCProviderSettings
+        <ProviderSettingsItem
           title="Microsoft"
           apiRef={microsoftAuthApiRef}
           icon={Star}
         />
       )}
       {providers.includes('github') && (
-        <OAuthProviderSettings
+        <ProviderSettingsItem
           title="Github"
           apiRef={githubAuthApiRef}
           icon={Star}
         />
       )}
       {providers.includes('gitlab') && (
-        <OAuthProviderSettings
+        <ProviderSettingsItem
           title="Gitlab"
           apiRef={gitlabAuthApiRef}
           icon={Star}
         />
       )}
       {providers.includes('okta') && (
-        <OIDCProviderSettings
+        <ProviderSettingsItem
           title="Okta"
           apiRef={oktaAuthApiRef}
           icon={Star}
         />
       )}
       {providers.includes('saml') && (
-        <OIDCProviderSettings
+        <ProviderSettingsItem
           title="SAML"
           apiRef={samlAuthApiRef}
           icon={Star}
         />
       )}
       {providers.includes('oauth2') && (
-        <OIDCProviderSettings
+        <ProviderSettingsItem
           title="YourOrg"
           apiRef={oauth2ApiRef}
           icon={Star}