+++ packages/core/src/layout/SignInPage/types.ts
@@ -20,20 +20,16 @@ import {
   SignInResult,
   ApiHolder,
   ApiRef,
-  OAuthApi,
   ProfileInfoApi,
   BackstageIdentityApi,
-  SessionStateApi,
-  SamlApi,
+  SessionApi,
 } from '@backstage/core-api';
 
 export type SignInConfig = {
   id: string;
   title: string;
   message: string;
-  apiRef:
-    | ApiRef<OAuthApi & ProfileInfoApi & BackstageIdentityApi & SessionStateApi>
-    | ApiRef<SamlApi>;
+  apiRef: ApiRef<ProfileInfoApi & BackstageIdentityApi & SessionApi>;
 };
 
 export type IdentityProviders = ('guest' | 'custom' | SignInConfig)[];
@@ -44,9 +40,7 @@ export type ProviderComponent = ComponentType<
 
 export type ProviderLoader = (
   apis: ApiHolder,
-  apiRef:
-    | ApiRef<OAuthApi & ProfileInfoApi & BackstageIdentityApi & SessionStateApi>
-    | ApiRef<SamlApi>,
+  apiRef: ApiRef<ProfileInfoApi & BackstageIdentityApi & SessionApi>,
 ) => Promise<SignInResult | undefined>;
 
 export type SignInProvider = {