+++ packages/create-app/templates/default-app/packages/backend/src/plugins/scaffolder.ts
@@ -10,7 +10,7 @@ import {
   GitlabPublisher,
   CreateReactAppTemplater,
   Templaters,
-  RepoVisilityOptions,
+  RepoVisibilityOptions,
 } from '@backstage/plugin-scaffolder-backend';
 import { Octokit } from '@octokit/rest';
 import { Gitlab } from '@gitbeaker/node';
@@ -42,7 +42,7 @@ export default async function createPlugin({
   const githubToken = config.getString('scaffolder.github.token');
   const repoVisibility = config.getString(
     'scaffolder.github.visibility',
-  ) as RepoVisilityOptions;
+  ) as RepoVisibilityOptions;
 
   const githubClient = new Octokit({ auth: githubToken });
   const githubPublisher = new GithubPublisher({